prevent a deadlock
New-plan: it's not worth the effort to have one 'control channel'; it
would make everything *a lot* easier if we just used a separate channel
for each control command.

This isn't as bad as it sounds at first.  Within a single thread, the
various functions (e.g. queueDeclare, exchangeDelete) are synchronous,
so, the library will only run as many commands concurrently as there are
threads in the client application >:).

It will suck on the rabbit side a bit, though.  Rabbit will start-up
about two Erlang processes for each channel.  Still, none of those
processes will interact in any way with the rest of the broker and won't
do IO or other expensive operations.  So, it'll just have to deal with
the overhead of setting up two new gen_server2's.

Also, it's not as if the control commands have any hard real-time limits.