tweaks to compile on 7.4 and... some other stuff
Added use of BangPatterns for GHC 7.4.

I have no idea what most of the commited stuff is.  Serves me right for
leaving an dirty head.

Regardless, the plan now is to junk all the /abstract/ messagin stuff
and provide a working bare-bones AMQP library in the spirit of the
RabbitMQ clients.  We'll be able to do any further abstraction on top of
that, if we have time.

I also think we're running into a TH bug with 7.4 here:
  http://hackage.haskell.org/trac/ghc/ticket/5375

This needs more thought, though.